Revert "RTS configurable from jenkins"
