+++ HpToolsLauncher/HpToolsLauncher.csproj
@@ -116,12 +116,9 @@
     <Compile Include="TestParameterInfo.cs" />
     <Compile Include="TestRunners\ParallelTestRunner.cs" />
     <Compile Include="TestRunners\PerformanceTestRunner.cs" />
-    <Compile Include="SummaryDataLogger.cs" />
+    <Compile Include="TestRunners\SummaryDataLogger.cs" />
     <Compile Include="TestRunResults.cs" />
     <Compile Include="TestRunners\GuiTestRunner.cs" />
-    <Compile Include="RTS\AdditionalAttributeModel.cs" />
-    <Compile Include="RTS\RTSHelper.cs" />
-    <Compile Include="RTS\ScriptRTSModel.cs" />
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DesignTime>True</DesignTime>
