+++ src/main/resources/com/microfocus/application/automation/tools/lr/Messages.properties
@@ -20,6 +20,3 @@
 
 ProductName=LoadRunner
 SummaryDataLogModel=Summary Data Log Model
-ScriptRTSSetModel=Script RTS Set Model
-ScriptRTSModel=Script RTS Model
-AdditionalAttributeModel=Additional Attribute Model
