--- src/main/resources/com/microfocus/application/automation/tools/lr/model/ScriptRTSModel/config.jelly
@@ -1,33 +1,0 @@
-<!--
-  ~
-  ~  Certain versions of software and/or documents (“Material”) accessible here may contain branding from
-  ~  Hewlett-Packard Company (now HP Inc.) and Hewlett Packard Enterprise Company.  As of September 1, 2017,
-  ~  the Material is now offered by Micro Focus, a separately owned and operated company.  Any reference to the HP
-  ~  and Hewlett Packard Enterprise/HPE marks is historical in nature, and the HP and Hewlett Packard Enterprise/HPE
-  ~  marks are the property of their respective owners.
-  ~ __________________________________________________________________
-  ~ MIT License
-  ~
-  ~ © Copyright 2012-2018 Micro Focus or one of its affiliates.
-  ~
-  ~ The only warranties for products and services of Micro Focus and its affiliates
-  ~ and licensors (“Micro Focus”) are set forth in the express warranty statements
-  ~ accompanying such products and services. Nothing herein should be construed as
-  ~ constituting an additional warranty. Micro Focus shall not be liable for technical
-  ~ or editorial errors or omissions contained herein.
-  ~ The information contained herein is subject to change without notice.
-  ~ ___________________________________________________________________
-  ~
-  -->
-<?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
-
-  <H3>Script</H3>
-  <f:entry title="${%ScriptName}" field="scriptName">
-    <f:textbox field="scriptName"/>
-  </f:entry>
-  <f:block>
-      <f:repeatableProperty field="additionalAttributes" minimum="1" add="+ Additional Attribute"/>
-  </f:block>
-  <input type="button" value="X" class="repeatable-delete danger" />
-</j:jelly>
