+++ src/main/resources/com/microfocus/application/automation/tools/lr/model/SummaryDataLogModel/config.jelly
@@ -20,7 +20,9 @@
   ~
   -->
 <?jelly escape-by-default='true'?>
-<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout"
+         xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:cv="/lib/custom">
+
     <f:entry title="${%PollingInterval}" field="pollingInterval">
         <f:textbox name="pollingInterval" value="${instance.pollingInterval}" default=""/>
     </f:entry>
