+++ src/main/resources/com/microfocus/application/automation/tools/pipelineSteps/LoadRunnerTestStep/config.jelly
@@ -63,11 +63,6 @@
                     <f:property field="summaryDataLogModel" />
                 </f:entry>
             </f:section>
-            <f:section title="${%RuntimeSettings}">
-                <f:entry name="Runtime Settings" field="scriptRTSSetModel">
-                    <f:property field="scriptRTSSetModel"/>
-                </f:entry>
-            </f:section>
         </f:advanced>
     </f:section>
 </j:jelly>
