+++ src/main/resources/com/microfocus/application/automation/tools/pipelineSteps/LoadRunnerTestStep/config.properties
@@ -24,4 +24,3 @@ DontForgetThePublisher=Make sure to enable the <strong>Publish Micro Focus \
     tests result</strong> option in the <strong>Post-build \
     Actions</strong> section. This allows the tests results to be published.
 SummaryDataLog=Summary Data Log
-RuntimeSettings=Runtime Settings
